+++ src/components/Banner/SiteBanner/index.tsx
@@ -8,7 +8,7 @@ import { theme } from '../../../theme/docsTheme';
 import { isBrowser } from '../../../utils/is-browser';
 import { normalizePath } from '../../../utils/normalize-path';
 import { fetchBanner } from '../../../utils/realm';
-import { SiteBannerContent } from '../../../types';
+import { SiteBannerContent } from './types';
 import BrandingShape from './BrandingShape';
 
 const getBannerSource = (src?: string) => {
