+++ src/components/Banner/SiteBanner/types.ts
